feat(server): updated models. Renamed metrics to Summary & added Metric

+++ server/types/workout.go
@@ -9,8 +9,8 @@ type WorkoutResponse struct {
 }
 
 type WorkoutData struct {
-	KcalBurned   int    `json:"kcalBurned"`
-	ActivityType string `json:"activityType"`
-	Name         string `json:"name"`
-	Duration     int    `json:"duration"`
+	KcalBurned   int     `json:"kcalBurned"`
+	ActivityType string  `json:"activityType"`
+	Name         string  `json:"name"`
+	Duration     float64 `json:"duration"`
 }
